+++ src/pages/Boards/BoardBar/BoardBar.jsx
@@ -37,7 +37,6 @@ function BoardBar() {
         gap: 1,
         paddingX: 2,
         overflowX: 'auto',
-        borderBottom: '1px solid white',
         backgroundColor: (theme) => (theme.palette.mode === 'dark' ? '#34495e' : '#1976d2'),
       }}
     >
