+++ src/pages/Boards/BoardContent/ListColumns/Column/ListCards/ListCards.jsx
@@ -1,0 +1,44 @@
+import { Box } from '@mui/material'
+import Card from './Card/Card'
+
+function ListCard() {
+  return (
+    <Box sx={{
+        p: '0 5px',
+        m: '0 5px',
+        display:'flex',
+        flexDirection:'column',
+        gap:1,
+        overflowX: 'hidden',
+        overflowy: 'auto',
+        maxHeight: (theme) => `calc(
+            ${theme.trello.BoardContentHeight} - 
+            ${theme.spacing(5)} -
+            ${theme.trello.columnHeaderHeight} -
+            ${theme.trello.columnFooterHeight}
+            )`,
+            '&::-webkit-scrollbar-thumb': {backgroundColor: '#ced0da',},
+            '&::-webkit-scrollbar-thumb:hover': {backgroundColor: '#bfc2cf'}
+        }}>
+            <Card/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+            <Card temporaryHideMedia/>
+        </Box>
+  )
+}
+
+export default ListCard
