+++ src/theme.js
@@ -1,17 +1,22 @@
 import { Height } from '@mui/icons-material'
 import { experimental_extendTheme as extendTheme } from '@mui/material/styles'
-import BoardContent from './pages/Boards/BoardContent'
+import BoardContent from './pages/Boards/BoardContent/BoardContent'
+import Column from './pages/Boards/BoardContent/ListColumns/Column/Column'
 
 const APP_BAR_HEIGHT = '58px'
 const BOARD_BAR_HEIGHT = '60px'
 const BOARD_CONTENT_HEIGHT = `calc(100vh - ${APP_BAR_HEIGHT} + ${BOARD_BAR_HEIGHT})`
+const COLUMN_HEADER_HEIGHT = '50px'
+const COLUMN_FOOTER_HEIGHT = '56px'
 
 // Create a theme instance.
 const theme = extendTheme({
   trello: {
     AppBarHeight: APP_BAR_HEIGHT,
     BoardBarHeight: BOARD_BAR_HEIGHT,
-    BoardContentHeight : BOARD_CONTENT_HEIGHT
+    BoardContentHeight : BOARD_CONTENT_HEIGHT,
+    columnHeaderHeight: COLUMN_FOOTER_HEIGHT,
+    columnFooterHeight: COLUMN_FOOTER_HEIGHT,
   },
   colorSchemes: {
     light: {},
